Seperate Socket.IO events.
