+++ src/interfaces/events/s2c.ts
@@ -1,0 +1,8 @@
+// ******************** //
+// Interfaces for the server to client events of Socket.IO
+// ******************** //
+
+export interface ServerToClientEvents {
+    // TODO: Start adding events which are called on their own here
+    // with io.emit(...), not callback-able
+}
