+++ src/main/server.ts
@@ -24,8 +24,10 @@ import fs from 'fs';
 import * as variables from 'other/variables';
 import { decideBooleanEnvValue } from 'other/utilities';
 import { Server } from 'socket.io';
-import { ClientToServerEvents, InterServerEvents, ServerToClientEvents } from 'interfaces/all';
 import { AnyError, MongoClient } from 'mongodb';
+import { ClientToServerEvents } from 'interfaces/events/c2s';
+import { ServerToClientEvents } from 'interfaces/events/s2c';
+import { InterServerEvents } from 'interfaces/events/inter';
 
 // Variables
 let io: Server<ClientToServerEvents, ServerToClientEvents, InterServerEvents>;
